Move IEnumOLEVERB to interop
